sr(api): move TrainingTriggerForAdmin model in devcomp
