+++ api/tests/devcomp/integration/infrastructure/repositories/training-repository_test.js
@@ -12,7 +12,7 @@ import { NotFoundError } from '../../../../../lib/domain/errors.js';
 import { TrainingSummary } from '../../../../../src/devcomp/domain/read-models/TrainingSummary.js';
 import { Training } from '../../../../../lib/domain/models/Training.js';
 import { UserRecommendedTraining } from '../../../../../lib/domain/read-models/UserRecommendedTraining.js';
-import { TrainingTriggerForAdmin } from '../../../../../lib/domain/read-models/TrainingTriggerForAdmin.js';
+import { TrainingTriggerForAdmin } from '../../../../../src/devcomp/domain/read-models/TrainingTriggerForAdmin.js';
 import { TrainingForAdmin } from '../../../../../src/devcomp/domain/read-models/TrainingForAdmin.js';
 import { TrainingTrigger } from '../../../../../lib/domain/models/TrainingTrigger.js';
 import { TrainingTriggerTube } from '../../../../../lib/domain/models/TrainingTriggerTube.js';
