+++ api/tests/devcomp/integration/infrastructure/repositories/training-trigger-repository_test.js
@@ -11,7 +11,7 @@ import {
 
 import * as trainingTriggerRepository from '../../../../../src/devcomp/infrastructure/repositories/training-trigger-repository.js';
 import { TrainingTrigger, TrainingTriggerTube } from '../../../../../lib/domain/models/index.js';
-import { TrainingTriggerForAdmin } from '../../../../../lib/domain/read-models/TrainingTriggerForAdmin.js';
+import { TrainingTriggerForAdmin } from '../../../../../src/devcomp/domain/read-models/TrainingTriggerForAdmin.js';
 import _ from 'lodash';
 import { NotFoundError } from '../../../../../lib/domain/errors.js';
 import { logger } from '../../../../../lib/infrastructure/logger.js';
