+++ api/tests/tooling/domain-builder/factory/build-training-trigger-for-admin.js
@@ -1,5 +1,5 @@
 import { buildTube } from './build-tube.js';
-import { TrainingTriggerForAdmin } from '../../../../lib/domain/read-models/TrainingTriggerForAdmin.js';
+import { TrainingTriggerForAdmin } from '../../../../src/devcomp/domain/read-models/TrainingTriggerForAdmin.js';
 import { TrainingTriggerTube } from '../../../../lib/domain/models/TrainingTriggerTube.js';
 
 const buildTrainingTriggerForAdmin = function ({
